requirements.txt
# UI
streamlit==1.33.0

# Core libs
numpy==1.24.4
pillow==9.5.0
opencv-python==4.9.0.80
pdf2image==1.16.3

# OCR
easyocr==1.7.1

# Transformers + PyTorch (CPU)
torch==2.0.1
torchvision==0.15.2
transformers==4.34.0

# Utilities
rapidfuzz==3.6.2
python-multipart==0.0.6
